intial code cleanup
