+++ .gitignore
@@ -1,0 +1,3 @@
+untracked
+.venv
+__pycache__
